+++ inc/staffSvg/index.ts
@@ -82,7 +82,6 @@ const createSheetDocumentFromSvgs = (svgs: string[], ly: string, lilyDocument: L
 	const attributes = lilyDocument.globalAttributes({readonly: true}) as LilyDocumentAttributeReadOnly;
 
 	const source = new TextSource(ly);
-	//const tieLocations = lilyDocument.getTiedNoteLocations(source).reduce((table, loc) => ((table[`${loc[0]}:${loc[1]}`] = true), table), {});
 	const tieLocations = lilyDocument.getTiedNoteLocations2().reduce((table, loc) => ((table[`${loc[0]}:${loc[1]}`] = true), table), {});
 	//logger.append("tieLocations:", Object.keys(tieLocations));
 
@@ -105,7 +104,8 @@ const postProcessSheetDocument = (sheet: SheetDocument, lilyDocument: LilyDocume
 	// align token id
 	const interpreter = lilyDocument.interpret();
 	const lilyNotation = interpreter.getNotation();
-	sheet.alignTokensWithNotation(lilyNotation);
+	if (lilyNotation)
+		sheet.alignTokensWithNotation(lilyNotation);
 
 	const attributes = lilyDocument.globalAttributes({readonly: true}) as LilyDocumentAttributeReadOnly;